Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
playwright
pandas
supabase
jinja2
python-multipart
python-dotenv
